+++ src/App.js
@@ -1,12 +1,19 @@
 import Main from './pages/Main';
 import './reset.css';
+import { styled } from 'styled-components';
 
 function App() {
   return (
     <div>
       <Main />
+      {/* <MainWrapper>TEST</MainWrapper> */}
     </div>
   );
 }
 
 export default App;
+
+// const MainWrapper = styled.div`
+//   width: 1400px;
+//   margin: 0 auto;
+// `;
