--- src/components/Footer.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Footer() {
-  return <div>Footer</div>;
-}
-
-export default Footer;
