--- src/components/Frame/Frame.jsx
@@ -1,1 +1,0 @@
-105,109,112,111,114,116,32,82,101,97,99,116,32,102,114,111,109,32,39,114,101,97,99,116,39,59,13,10,13,10,102,117,110,99,116,105,111,110,32,70,114,97,109,101,40,41,32,123,13,10,32,32,114,101,116,117,114,110,32,60,100,105,118,62,70,114,97,109,101,60,47,100,105,118,62,59,13,10,125,13,10,13,10,101,120,112,111,114,116,32,100,101,102,97,117,108,116,32,70,114,97,109,101,59,13,10
