+++ src/pages/DetailPage.jsx
